DESCW-1097 Report - Indivicual Project Reports > Project Status Summary
