+++ backend/src/models/reports/Tab_18_rpt_P_StatusSummary.js
@@ -1,17 +1,4 @@
 const dbConnection = require("@database/databaseConnection");
 const { knex } = dbConnection();
-
-/**
- * Gets data for the Divisional Project Reports - Project Dashboard report.
- *
- * @returns {any[]}
- */
-const Tab_18_rpt_P_StatusSummary = () => {
-  const query = knex().raw();
-
-  return query;
-};
-
-module.exports = {
-  Tab_18_rpt_P_StatusSummary,
-};
+// Because all of the queries are coming from the global project, we can do this.
+module.exports = require("@models/reports/project");
